refactor json parser
