--- scrapr/client/src/cheerioObject.json
@@ -1,1 +1,0 @@
-[object Object]
